Control.Biegunka

Description

Biegunka - configuration development library

Interpreters control

biegunka

Arguments

:: (Settings () -> Settings ())
User defined settings

-> Interpreter
Combined interpreters

-> Script Sources ()
Script to interpret

-> IO ExitCode

Entry point into the library

data Settings a

Settings common for all interpreters and also specific for this one

root :: HasRoot s => Lens' s FilePath

Biegunka root

appData :: Lens' (Settings a) FilePath

Biegunka profile files

colors :: Lens' (Settings a) ColorScheme

Pretty printing

data Templates

Wraps templating system, hiding implementation

Constructors

forall t . TemplateSystem t => Templates t

templates :: Lens' (Settings a) Templates

Templates mapping

Interpreters

data Interpreter

Abstract Interpreter data type

pause :: Interpreter

Interpreter that just waits user to press any key

confirm :: Interpreter

Interpreter that awaits user confirmation

dryRun :: Interpreter

Dry run interpreter

run :: Interpreter

Real run interpreter

check :: Interpreter

Check interpreter

Types

data Script s a

Newtype used to provide better error messages for type errors in DSL (for users, mostly)

data Scope

Language terms scopes [kind]

Constructors

Actions

Sources

Sources layer primitives

class Sourceable s where

Common Sources structure

Methods

actions :: Lens' s (Script Actions ())

Actions to run after source update

(==>) :: String -> FilePath -> s -> Script Sources ()

Actions layer primitives

link :: FilePath -> FilePath -> Script Actions ()

Links given file to specified filepath

 git "https://example.com/source.git" "git/source" $
   link "some-file" "anywhere"

Links ~/git/source/some-file to ~/anywhere.

register :: FilePath -> Script Actions ()

Links source to specified filepath

 git "https://example.com/source.git" "git/source" $
   register "somewhere"

Links ~/git/source to ~/somewhere.

copy :: FilePath -> FilePath -> Script Actions ()

Copies file or directory to specified filepath

 git "https://example.com/source.git" "git/source" $
   copy "some-file" "anywhere"

Copies ~/git/source/some-file to ~/anywhere.

copyFile :: FilePath -> FilePath -> Script Actions ()

Copies only single file to specified filepath

 git "https://example.com/source.git" "git/source" $
   copy "some-file" "anywhere"

Copies ~/git/source/some-file to ~/anywhere.

copyDirectory :: FilePath -> FilePath -> Script Actions ()

Copies file or directory to specified filepath

 git "https://example.com/source.git" "git/source" $
   copy "some-file" "anywhere"

Copies ~/git/source/some-file to ~/anywhere.

substitute :: FilePath -> FilePath -> Script Actions ()

Substitutes templates in HStringTemplate syntax in given file and writes result to specified filepath

 git "https://example.com/source.git" "git/source" $
   substitute "some-file.template" "anywhere"

Copies ~/git/source/some-file.template to ~/anywhere.

Substitutes templates in ~/anywhere with values from templates part of Controls

patch :: FilePath -> FilePath -> PatchSpec -> Script Actions ()

Applies the patch given the PatchSpec

 git "https://example.com/source.git" "git/source" $
   patch "some-patch.patch" "anywhere" (def { reversely = True })

Applies ~/git/source/some-patch.patch to ~/anywhere reversely.

data PatchSpec

Patch settings

Constructors

PatchSpec

Fields

strip :: Int

How many leading slashes to strip

reversely :: Bool

Apply in reverse?

raw :: FilePath -> [String] -> Script Actions ()

Monomorphised interface to sh quasiquoter for those who do not like -XTemplateHaskell (or -XQuasiQuotes)

 git "https://example.com/source.git" "git/source" $
   raw "/bin/echo" ["-n", "hello"]

Prints "hello" to stdout

Script environment

root :: HasRoot s => Lens' s FilePath

Biegunka root

source :: HasSource s => Lens' s FilePath

Source root

Modifiers

profile :: String -> Script Sources a -> Script Sources a

Provides convenient Sources grouping. May be nested

Information about sources and files related to a particular profile profile could be found in ~/.biegunka/profiles/.

Example usage:

 profile "dotfiles" $ do
   group "mine" $
     git "https://github.com/supki/.dotfiles"
       ...
   group "not-mine" $
     git "https://github.com/dmalikov/dotfiles"
       ...
 profile "experimental" $ do
   git "https://github.com/ekmett/lens"
     ...

group :: String -> Script Sources a -> Script Sources a

Alias for profile. May be useful for nested grouping

role :: String -> Script Sources a -> Script Sources a

Alias for profile. Everyone uses roles for something

sudo :: User u -> Script s a -> Script s a

Change effective user id for wrapped commands

data User u where

User setting modifier

Constructors

UserID :: CUid -> User CUid

Username :: String -> User String

retries :: Int -> Script s a -> Script s a

Change maximum retries count

reacting :: React -> Script s a -> Script s a

Change reaction pattern when retries are all failed

data React

Failure reaction

Used then all retries errored

Constructors

Ignorant

Abortive

prerequisiteOf :: Script Sources a -> Script Sources b -> Script Sources b

Execute scripts sequentially Connects two scripts which forces them to run sequentially one after another.

(<~>) :: Script Sources a -> Script Sources b -> Script Sources b

Infix alias for prerequisiteOf

Auxiliary

into :: FilePath -> FilePath

A hack to support the notion of making destination FilePath inside some directory

Options autogeneration

class Typeable a => Data a

The Data class comprehends a fundamental primitive gfoldl for folding over constructor applications, say terms. This primitive can be instantiated in several ways to map over the immediate subterms of a term; see the gmap combinators later in this class. Indeed, a generic programmer does not necessarily need to use the ingenious gfoldl primitive but rather the intuitive gmap combinators. The gfoldl primitive is completed by means to query top-level constructors, to turn constructor representations into proper terms, and to list all possible datatype constructors. This completion allows us to serve generic programming scenarios like read, show, equality, term generation.

The combinators gmapT, gmapQ, gmapM, etc are all provided with default definitions in terms of gfoldl, leaving open the opportunity to provide datatype-specific definitions. (The inclusion of the gmap combinators as members of class Data allows the programmer or the compiler to derive specialised, and maybe more efficient code per datatype. Note: gfoldl is more higher-order than the gmap combinators. This is subject to ongoing benchmarking experiments. It might turn out that the gmap combinators will be moved out of the class Data.)

Conceptually, the definition of the gmap combinators in terms of the primitive gfoldl requires the identification of the gfoldl function arguments. Technically, we also need to identify the type constructor c for the construction of the result type from the folded term type.

In the definition of gmapQx combinators, we use phantom type constructors for the c in the type of gfoldl because the result type of a query does not involve the (polymorphic) type of the term argument. In the definition of gmapQl we simply use the plain constant type constructor because gfoldl is left-associative anyway and so it is readily suited to fold a left-associative binary operation over the immediate subterms. In the definition of gmapQr, extra effort is needed. We use a higher-order accumulation trick to mediate between left-associative constructor application vs. right-associative binary operation (e.g., (:)). When the query is meant to compute a value of type r, then the result type withing generic folding is r -> r. So the result of folding is a function to which we finally pass the right unit.

With the -XDeriveDataTypeable option, GHC can generate instances of the Data class automatically. For example, given the declaration

 data T a b = C1 a b | C2 deriving (Typeable, Data)

GHC will generate an instance that is equivalent to

 instance (Data a, Data b) => Data (T a b) where
     gfoldl k z (C1 a b) = z C1 `k` a `k` b
     gfoldl k z C2       = z C2

     gunfold k z c = case constrIndex c of
                         1 -> k (k (z C1))
                         2 -> z C2

     toConstr (C1 _ _) = con_C1
     toConstr C2       = con_C2

     dataTypeOf _ = ty_T

 con_C1 = mkConstr ty_T "C1" [] Prefix
 con_C2 = mkConstr ty_T "C2" [] Prefix
 ty_T   = mkDataType "Module.T" [con_C1, con_C2]

This is suitable for datatypes that are exported transparently.

class Typeable a

The class Typeable allows a concrete representation of a type to be calculated.

options :: Data a => [a] -> IO (a, Runner b)

Run constructed parser

Quasiquoters

multiline :: QuasiQuoter

QuasiQuoter for raw multiline strings

sh :: QuasiQuoter

shell :: FilePath -> QuasiQuoter

Settings

Colors

data ColorScheme

Colors used in logger

Constructors

ColorScheme

Fields

_actionColor :: Doc -> Doc

_sourceColor :: Doc -> Doc

_srcColor :: Doc -> Doc

_dstColor :: Doc -> Doc

_errorColor :: Doc -> Doc

_retryColor :: Doc -> Doc

actionColor :: Lens' ColorScheme (Doc -> Doc)

Action color

sourceColor :: Lens' ColorScheme (Doc -> Doc)

Source color

srcColor :: Lens' ColorScheme (Doc -> Doc)

Src color

dstColor :: Lens' ColorScheme (Doc -> Doc)

Dst color

errorColor :: Lens' ColorScheme (Doc -> Doc)

Error color

retryColor :: Lens' ColorScheme (Doc -> Doc)

Retry color

noColors :: ColorScheme

Disable colors

Mode

mode :: Lens' (Settings a) Mode

Biegunka mode

data Mode

Constructors

Offline

Online

Little helpers

(~>) :: a -> b -> (a, b)

An alias for '(,)' for better looking pairing