Control.Biegunka.Biegunka

Wrap/unwrap biegunka interpreters

biegunka

Arguments

:: (Settings () -> Settings ())
User defined settings

-> Interpreter
Combined interpreters

-> Script Sources ()
Script to interpret

-> IO ExitCode

Entry point into the library

data Interpreter

Abstract Interpreter data type

interpret :: (Settings () -> Free (Term Annotate Sources) () -> IO ExitCode -> IO ExitCode) -> Interpreter

Construct Interpreter

Provides Settings, script and continuation, which is supposed to be called on success to interpreter

interpretOptimistically :: (Settings () -> Free (Term Annotate Sources) () -> IO ()) -> Interpreter

Construct Interpreter optimistically

It is optimistic in a sense what it always calls the continuation, provided that no exceptions were thrown

Generic interpreters

pause :: Interpreter

Interpreter that just waits user to press any key

confirm :: Interpreter

Interpreter that awaits user confirmation

Auxiliary

expandHome :: String -> IO String

Expand "~" at the start of pattern