Control.Biegunka.Execute.Watcher

Description

Execution Watcher

Watcher

data Watcher

Watcher abstract data type

Watcher watches number of jobs currently running and subtasks ids that are already done execution

new :: IO Watcher

Create new Watcher

Newly created Watcher does not know about any jobs and tasks

wait :: Watcher -> IO ()

Create new Watcher

Waits until job counter reaches 0

 new >>= wait ≡ return ()

register :: MonadIO m => Watcher -> m ()

Watcher jobs counter goes up

unregister :: MonadIO m => Watcher -> m ()

Watcher jobs counter goes down

done :: MonadIO m => Watcher -> Token -> m ()

notify Watcher subtask is done

waitDone :: MonadIO m => Watcher -> Set Token -> m ()

Wait until all those subtasks are done