Control.Biegunka.Groups

Description

Groups data control

Uses acid-state to store data before script invocations.

Interpreters can modify (create/update/delete) data any way they want:

First, interpreter must call open to get data handle

Next, interpreter changes these and call commit

Last, interpreter says close

Documentation

data Partitioned a

Groups data state

Consists of acid-state handle and two parts: relevant groups and irrelevant groups

Relevant groups (or these) are groups mentioned in script so interpreter won't deal with others (or those)

data Groups

All groups data

A mapping from group names to GroupRecords

newtype GroupRecord

Group data record

A mapping from Source data to set of FileRecords

Constructors

GR

Fields

unGR :: Map SourceRecord (Set FileRecord)

data SourceRecord

Source data record

Constructors

SR

Fields

sourceType :: String

fromLocation :: FilePath

sourcePath :: FilePath

sourceOwner :: Maybe (Either String Int)

data FileRecord

File data record

Constructors

FR

Fields

fileType :: String

fromSource :: FilePath

filePath :: FilePath

fileOwner :: Maybe (Either String Int)

these :: Lens' (Partitioned a) a

Groups targeted by script

those :: Lens' (Partitioned a) a

All other groups

groups :: Lens' Groups (Map String GroupRecord)

All groups data

open :: Settings () -> IO (Partitioned Groups)

Open groups data from disk

Searches appData/groups path for groups data. Starts empty if nothing is found

commit :: Partitioned Groups -> IO ()

Update groups data

Combines these and those to get full state

close :: Partitioned Groups -> IO ()

Save groups data to disk

fromScript :: Free (Term Annotate Sources) a -> Groups

Extract groups data from script

Won't get all mentioned groups but only those for which there is some useful action to do.

diff :: Eq b => (a -> [b]) -> a -> a -> [b]

Get groups difference

files :: Groups -> [FilePath]

Get all destination filepaths in Groups

sources :: Groups -> [FilePath]

Get all sources location in Groups

who :: Maybe (Either String Int) -> String