Control.Biegunka.Options

Documentation

options :: Data a => [a] -> IO (a, Runner b)

Run constructed parser

parser :: Data a => [a] -> Parser (a, Runner b)

Create a parser for a list of data constructors, interpreters, and modes

constructorOptions :: Data a => [a] -> Parser a

Create a parser for a list of data constructors

constructorOption :: Data a => Map String Int -> a -> (Map String Int, Parser a)

Create a parser for a data constructor

constructorName :: Data a => a -> String

Get transformed constructor name

transformConstructor :: String -> String

Transform data constructor name

>>> transformcCnstructor "FooBarBaz"
"foo-bar-baz"