Control.Biegunka.QQ

Description

QuasiQuoters

Documentation

sh :: QuasiQuoter

QuasiQuoter for shell scripts

class Eval r where

Different interesting return types for sh QuasiQuoter

Instances mostly resemble the types of things in System.Process

Methods

eval :: String -> [String] -> r

class Embed a where

Embed haskell values into shell scripts

Instances provided for all Prelude data types for which it makes sense

I recommend using -XExtendedDefaultRules for modules where you want to embed values, it would save for annoying type annotations for numeric literals

Methods

embed :: a -> String