Control.Biegunka.Script

Description

Configuration script machinery

Gets interesting static information from script before passing it to interpreters, which then make use of it

Script types

newtype Script s a

Newtype used to provide better error messages for type errors in DSL (for users, mostly)

Constructors

Script

Fields

unScript :: ReaderT AnnotationsEnv (StateT AnnotationsState (Free (Term Annotate s))) a

data family Annotate sc :: *

Language Term annotation depending on their Scope

Implementation details

data AnnotationsState

Script construction state

data AnnotationsEnv

Get annotated script

runScript :: AnnotationsState -> AnnotationsEnv -> Script s a -> Free (Term Annotate s) (a, AnnotationsState)

Get annotated DSL and resulting state

runScript' :: AnnotationsState -> AnnotationsEnv -> Script s a -> (Free (Term Annotate s) a, AnnotationsState)

Get annotated DSL and resulting state

evalScript :: AnnotationsState -> AnnotationsEnv -> Script s a -> Free (Term Annotate s) a

Get annotated DSL

Script mangling

script :: Term Annotate s a -> Script s a

Lift DSL term to the Script

annotate :: Script s a -> ReaderT AnnotationsEnv (StateT AnnotationsState (Free (Term Annotate t))) (Free (Term Annotate s) a)

Annotate DSL

sourced :: String -> URI -> FilePath -> Script Actions () -> (FilePath -> IO ()) -> Script Sources ()

Abstract away all plumbing needed to make source

actioned :: (FilePath -> FilePath -> Action) -> Script Actions ()

Get Actions scope script from FilePath mangling

constructDestinationFilepath :: FilePath -> FilePath -> FilePath -> FilePath

Construct destination FilePath

>>> constructDestinationFilepath "" "" ""
""

>>> constructDestinationFilepath "/root" "from" "to"
"/root/to"

>>> constructDestinationFilepath "/root" "from" "/to"
"/to"

>>> constructDestinationFilepath "/root" "from" "to/"
"/root/to/from"

>>> constructDestinationFilepath "/root" "from" "/to/"
"/to/from"

Lenses

app :: Lens' AnnotationsEnv FilePath

Biegunka filepath root

profileName :: Lens' AnnotationsEnv String

Current profile name

sourcePath :: Lens' AnnotationsEnv FilePath

Current source filepath

sourceURL :: Lens' AnnotationsEnv String

Current source url

profiles :: Lens' AnnotationsState (Set String)

All profiles encountered so far

token :: Lens' AnnotationsState Int

Unique token for each 'TP'/'TS'

order :: Lens' AnnotationsState Int

Current action order

sourceReaction :: Lens' AnnotationsEnv React

How to react on source failure

actionReaction :: Lens' AnnotationsEnv React

How to react on action failure

activeUser :: Lens' AnnotationsEnv (Maybe UserW)

Current user

maxRetries :: Lens' AnnotationsEnv Retry

Maximum retries count

Misc

type URI = String

Repository URI (like git@github.com:whoever/whatever.git)

data UserW

Constructors

forall u . UserW (User u)

data User u where

User setting modifier

Constructors

UserID :: CUid -> User CUid

Username :: String -> User String

data React

Failure reaction

Constructors

Ignorant

Abortive

newtype Retry

Constructors

Retry

Fields

unRetry :: Int

incr :: Retry -> Retry