Control.Biegunka.Script.Token

Description

Infinite stream of unique tokens monad transformer

Types

Monad

data StreamT e m a

Monad transformer for the never ending stream of stuff

mtl-style class

class Monad m => MonadStream e m | m -> e where

mtl-style class, to avoid manual lifting

Methods

next :: m e

Get next element from the stream; advance that stream by one step

peek :: m e

Take a look at the next stream element, without modifying the stream

Token stream

data Infinite a

Never ending stream of stuff

Constructors

a :< (Infinite a)

data Token

Token.

Supposedly, the only way to get tokens is using tokens, so they are somewhat unique

Do things with StreamT

runStreamT :: Monad m => Infinite e -> StreamT e m a -> m a

Run StreamT with the supplied stream.

>>> runStreamT (fromList (map Token [1..])) (replicateM 3 next)
[Token 1,Token 2,Token 3]

mapStreamT :: (m (Infinite e, a) -> n (Infinite e, b)) -> StreamT e m a -> StreamT e n b

Map both the value and underlying monad

Tokens

tokens :: Infinite Token

Infinite stream of tokens

noTokens :: Infinite Void

Infinite stream of nothing, very zen

fromList :: [a] -> Infinite a

Get Infinite stream from list

>>> fromList [1,2,3]
1 :< (2 :< (3 :< *** Exception: Control.Biegunka.Script.Token.fromList: supplied list is not infinite