Control.Biegunka.Settings

Description

Controlling biegunka interpreters and their composition

Wrap/unwrap biegunka interpreters

biegunka

Arguments

:: (Settings () -> Settings ())
User defined settings

-> Interpreter
Combined interpreters

-> Script Sources ()
Script to interpret

-> IO ()

Common Interpreters Controls wrapper

newtype Interpreter

Interpreter newtype. Takes Controls, Script and performs some IO

Constructors

I

Fields

runInterpreter :: Settings () -> Free (Term Annotate Sources) () -> AnnotationsState -> IO () -> IO ()

interpret :: (Settings () -> Free (Term Annotate Sources) () -> AnnotationsState -> IO ()) -> Interpreter

Interpreter that calls its continuation after interpretation

Settings common for all interpreters

data Settings a

Settings common for all interpreters and also specific for this one

root :: Lens' (Settings a) FilePath

Root path for Source layer

appData :: Lens' (Settings a) FilePath

Biegunka profile files

logger :: Lens' (Settings a) (Doc -> IO ())

Logger channel

colors :: Lens' (Settings a) ColorScheme

Pretty printing

local :: Lens (Settings a) (Settings b) a b

Interpreter controls

templates :: Lens' (Settings a) Templates

Templates mapping

data Templates

Wraps templating system, hiding implementation

Constructors

forall t . TemplateSystem t => Templates t

Color scheme controls

data ColorScheme

Colors used in logger

Constructors

ColorScheme

Fields

_actionColor :: Doc -> Doc

_sourceColor :: Doc -> Doc

_srcColor :: Doc -> Doc

_dstColor :: Doc -> Doc

_errorColor :: Doc -> Doc

_retryColor :: Doc -> Doc

noColors :: ColorScheme

Disable colors

actionColor :: Lens' ColorScheme (Doc -> Doc)

Action color

sourceColor :: Lens' ColorScheme (Doc -> Doc)

Source color

srcColor :: Lens' ColorScheme (Doc -> Doc)

Src color

dstColor :: Lens' ColorScheme (Doc -> Doc)

Dst color

errorColor :: Lens' ColorScheme (Doc -> Doc)

Error color

retryColor :: Lens' ColorScheme (Doc -> Doc)

Retry color

Generic interpreters

pause :: Interpreter

Interpreter that just waits user to press any key

confirm :: Interpreter

Interpreter that awaits user confirmation