Control.Biegunka.Source.Git

Description

Support for git repositories as Sources

Source layer

git' :: Target p => URI -> p -> Git -> Script Sources ()

Clone repository from the url to the specified path using provided Git settings. Sample:

 git' "git@github.com:edwinb/Idris-dev" "git/Idris-dev" $ def
   & remotes .~ ["origin", "stream"]
   & branch .~ "develop"
   & actions .~ do
       link "contribs/tool-support/vim" ".vim/bundle/idris-vim"

Clone repository from https://github.com/edwinb/Idris-dev.git to ~/git/Idris-dev

Merge origin/develop into develop

Merge stream/develop into develop

Checkout to develop

Link ~/git/Idris-dev/contribs/tool-support/vim to ~/.vim/bundle/Idris-vim

git :: Target p => URI -> p -> Script Actions () -> Script Sources ()

Wrapper over git' that provides easy specification of actions field

git_ :: Target p => URI -> p -> Script Sources ()

Wrapper over git that does not provide anything

Types

data Git

Git repository's settings

Constructors

Git

Fields

gitactions :: Script Actions ()

Actions to run after repository update

_remotes :: [Remote]

Remotes to merge on update

_branch :: Branch

Branch to track

defaultGit :: Git

Do nothing except pulling origin/master into master

Lenses

actions :: Sourceable s => Lens' s (Script Actions ())

Actions to run after source update

remotes :: Lens' Git [Remote]

Remotes to merge on update

branch :: Lens' Git Branch

Branch to track

Type synonyms

Convenient self-described types to remind yourself where is which argument

type Branch = String

Branch name (like master or develop)

type Remote = String

Remote name (like origin or upstream)

type URI = String

Repository URI (like git@github.com:whoever/whatever.git)